requirements.txt
streamlit
pymupdf
python-docx
arabic-reshaper
python-bidi
qrcode
pillow